refactor: refactor autoOpenProejct, readme and collaborator handlers (#11881)
* refactor: refactor autoOpenProejct and readme handlers

* fix: fix CodeQL alert

* test: update ut
