+++ packages/vscode-extension/test/extension/handlers.test.ts
@@ -5,7 +5,6 @@ import {
   ConfigFolderName,
   FxError,
   Inputs,
-  ManifestUtil,
   OptionItem,
   Platform,
   Result,
@@ -17,16 +16,13 @@ import {
 } from "@microsoft/teamsfx-api";
 import {
   AppDefinition,
-  CollaborationState,
   DepsManager,
   DepsType,
   UnhandledError,
   UserCancelError,
   environmentManager,
   featureFlagManager,
-  manifestUtils,
   pathUtils,
-  pluginManifestUtils,
   teamsDevPortalClient,
 } from "@microsoft/teamsfx-core";
 import * as globalState from "@microsoft/teamsfx-core/build/common/globalState";
@@ -40,15 +36,14 @@ import * as uuid from "uuid";
 import * as vscode from "vscode";
 import { AzureAccountManager } from "../../src/commonlib/azureLogin";
 import { signedIn, signedOut } from "../../src/commonlib/common/constant";
-import VsCodeLogInstance, { VsCodeLogProvider } from "../../src/commonlib/log";
+import VsCodeLogInstance from "../../src/commonlib/log";
 import M365TokenInstance from "../../src/commonlib/m365Login";
-import { DeveloperPortalHomeLink, GlobalKey } from "../../src/constants";
+import { DeveloperPortalHomeLink } from "../../src/constants";
 import { PanelType } from "../../src/controls/PanelType";
 import { WebviewPanel } from "../../src/controls/webviewPanel";
 import * as debugConstants from "../../src/debug/common/debugConstants";
 import * as getStartedChecker from "../../src/debug/depsChecker/getStartedChecker";
 import * as launch from "../../src/debug/launch";
-import * as runIconHandlers from "../../src/debug/runIconHandler";
 import * as errorCommon from "../../src/error/common";
 import { ExtensionErrors } from "../../src/error/error";
 import { TreatmentVariableValue } from "../../src/exp/treatmentVariables";
@@ -66,8 +61,6 @@ import * as vsc_ui from "../../src/qm/vsc_ui";
 import { VsCodeUI } from "../../src/qm/vsc_ui";
 import { ExtTelemetry } from "../../src/telemetry/extTelemetry";
 import * as extTelemetryEvents from "../../src/telemetry/extTelemetryEvents";
-import { TelemetryEvent } from "../../src/telemetry/extTelemetryEvents";
-import * as appDefinitionUtils from "../../src/utils/appDefinitionUtils";
 import { updateAutoOpenGlobalKey } from "../../src/utils/globalStateUtils";
 import * as localizeUtils from "../../src/utils/localizeUtils";
 import * as migrationUtils from "../../src/utils/migrationUtils";
@@ -641,103 +634,6 @@ describe("handlers", () => {
     sandbox.assert.calledOnceWithExactly(createOrShow, PanelType.SampleGallery, []);
   });
 
-  it("openReadMeHandler", async () => {
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "isTeamsFxProject").value(true);
-    const executeCommands = sandbox.stub(vscode.commands, "executeCommand");
-    sandbox
-      .stub(vscode.workspace, "workspaceFolders")
-      .value([{ uri: { fsPath: "readmeTestFolder" } }]);
-    sandbox.stub(fs, "pathExists").resolves(true);
-    const openTextDocumentStub = sandbox
-      .stub(vscode.workspace, "openTextDocument")
-      .resolves({} as any as vscode.TextDocument);
-
-    await handlers.openReadMeHandler([extTelemetryEvents.TelemetryTriggerFrom.Auto]);
-
-    chai.assert.isTrue(openTextDocumentStub.calledOnce);
-    chai.assert.isTrue(executeCommands.calledOnce);
-  });
-
-  it("openReadMeHandler - create project", async () => {
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "isTeamsFxProject").value(false);
-    sandbox.stub(globalVariables, "core").value(undefined);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Yes",
-            run: (options as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-    await handlers.openReadMeHandler([extTelemetryEvents.TelemetryTriggerFrom.Auto]);
-
-    chai.assert.isTrue(showMessageStub.calledOnce);
-  });
-
-  it("openReadMeHandler - open folder", async () => {
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "isTeamsFxProject").value(false);
-    sandbox.stub(globalVariables, "core").value(undefined);
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Yes",
-            run: (items[0] as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-    await handlers.openReadMeHandler([extTelemetryEvents.TelemetryTriggerFrom.Auto]);
-
-    chai.assert.isTrue(executeCommandStub.calledOnce);
-  });
-
-  it("openReadMeHandler - function notification bot template", async () => {
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "isTeamsFxProject").value(true);
-    sandbox
-      .stub(vscode.workspace, "workspaceFolders")
-      .value([{ uri: { fsPath: "readmeTestFolder" } }]);
-    sandbox.stub(TreatmentVariableValue, "inProductDoc").value(true);
-    sandbox.stub(fs, "pathExists").resolves(true);
-    sandbox.stub(fs, "readFile").resolves(Buffer.from("## Get Started with the Notification bot"));
-    const createOrShow = sandbox.stub(WebviewPanel, "createOrShow");
-
-    await handlers.openReadMeHandler([extTelemetryEvents.TelemetryTriggerFrom.Auto]);
-
-    sandbox.assert.calledOnceWithExactly(
-      createOrShow,
-      PanelType.FunctionBasedNotificationBotReadme
-    );
-  });
-
-  it("openReadMeHandler - restify notification bot template", async () => {
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "isTeamsFxProject").value(true);
-    sandbox
-      .stub(vscode.workspace, "workspaceFolders")
-      .value([{ uri: { fsPath: "readmeTestFolder" } }]);
-    sandbox.stub(TreatmentVariableValue, "inProductDoc").value(true);
-    sandbox.stub(fs, "pathExists").resolves(true);
-    sandbox
-      .stub(fs, "readFile")
-      .resolves(Buffer.from("## Get Started with the Notification bot restify"));
-    const createOrShow = sandbox.stub(WebviewPanel, "createOrShow");
-
-    await handlers.openReadMeHandler([extTelemetryEvents.TelemetryTriggerFrom.Auto]);
-
-    sandbox.assert.calledOnceWithExactly(
-      createOrShow,
-      PanelType.RestifyServerNotificationBotReadme
-    );
-  });
-
   describe("decryptSecret", function () {
     const sandbox = sinon.createSandbox();
 
@@ -803,138 +699,6 @@ describe("handlers", () => {
       sinon.assert.calledOnce(showMessage);
       sinon.assert.calledOnce(sendTelemetryEvent);
       sinon.assert.calledOnce(sendTelemetryErrorEvent);
-    });
-  });
-
-  describe("permission v3", function () {
-    const sandbox = sinon.createSandbox();
-
-    this.afterEach(() => {
-      sandbox.restore();
-    });
-
-    it("happy path: grant permission", async () => {
-      sandbox.stub(globalVariables, "core").value(new MockCore());
-      sandbox.stub(vsc_ui, "VS_CODE_UI").value({
-        selectOption: () => Promise.resolve(ok({ type: "success", result: "grantPermission" })),
-      });
-      sandbox.stub(MockCore.prototype, "grantPermission").returns(
-        Promise.resolve(
-          ok({
-            state: CollaborationState.OK,
-            userInfo: {
-              userObjectId: "fake-user-object-id",
-              userPrincipalName: "fake-user-principle-name",
-            },
-            permissions: [
-              {
-                name: "name",
-                type: "type",
-                resourceId: "id",
-                roles: ["Owner"],
-              },
-            ],
-          })
-        )
-      );
-
-      const result = await handlers.manageCollaboratorHandler("env");
-      chai.expect(result.isOk()).equals(true);
-    });
-
-    it("happy path: list collaborator", async () => {
-      sandbox.stub(globalVariables, "core").value(new MockCore());
-      sandbox.stub(vsc_ui, "VS_CODE_UI").value({
-        selectOption: () => Promise.resolve(ok({ type: "success", result: "listCollaborator" })),
-      });
-      sandbox.stub(MockCore.prototype, "listCollaborator").returns(
-        Promise.resolve(
-          ok({
-            state: CollaborationState.OK,
-            collaborators: [
-              {
-                userPrincipalName: "userPrincipalName",
-                userObjectId: "userObjectId",
-                isAadOwner: true,
-                teamsAppResourceId: "teamsAppResourceId",
-              },
-            ],
-          })
-        )
-      );
-      const vscodeLogProviderInstance = VsCodeLogProvider.getInstance();
-      sandbox.stub(vscodeLogProviderInstance, "outputChannel").value({
-        name: "name",
-        append: (value: string) => {},
-        appendLine: (value: string) => {},
-        replace: (value: string) => {},
-        clear: () => {},
-        show: (...params: any[]) => {},
-        hide: () => {},
-        dispose: () => {},
-      });
-
-      const result = await handlers.manageCollaboratorHandler("env");
-      chai.expect(result.isOk()).equals(true);
-    });
-
-    it("happy path: list collaborator throws error", async () => {
-      sandbox.stub(globalVariables, "core").value(new MockCore());
-      sandbox.stub(vsc_ui, "VS_CODE_UI").value({
-        selectOption: () => Promise.resolve(ok({ type: "success", result: "listCollaborator" })),
-      });
-      sandbox.stub(MockCore.prototype, "listCollaborator").throws(new Error("Error"));
-      const vscodeLogProviderInstance = VsCodeLogProvider.getInstance();
-      sandbox.stub(vscodeLogProviderInstance, "outputChannel").value({
-        name: "name",
-        append: (value: string) => {},
-        appendLine: (value: string) => {},
-        replace: (value: string) => {},
-        clear: () => {},
-        show: (...params: any[]) => {},
-        hide: () => {},
-        dispose: () => {},
-      });
-
-      const result = await handlers.manageCollaboratorHandler("env");
-      chai.expect(result.isErr()).equals(true);
-    });
-
-    it("happy path: list collaborator throws login error", async () => {
-      sandbox.stub(globalVariables, "core").value(new MockCore());
-      sandbox.stub(vsc_ui, "VS_CODE_UI").value({
-        selectOption: () => Promise.resolve(ok({ type: "success", result: "listCollaborator" })),
-      });
-      const showErrorMessageStub = sandbox.stub(vscode.window, "showErrorMessage");
-      sandbox
-        .stub(MockCore.prototype, "listCollaborator")
-        .throws(new Error("Cannot get user login information"));
-      const vscodeLogProviderInstance = VsCodeLogProvider.getInstance();
-      sandbox.stub(vscodeLogProviderInstance, "outputChannel").value({
-        name: "name",
-        append: (value: string) => {},
-        appendLine: (value: string) => {},
-        replace: (value: string) => {},
-        clear: () => {},
-        show: (...params: any[]) => {},
-        hide: () => {},
-        dispose: () => {},
-      });
-
-      const result = await handlers.manageCollaboratorHandler("env");
-      chai.expect(result.isErr()).equals(true);
-      chai.assert.isTrue(showErrorMessageStub.called);
-    });
-
-    it("User Cancel", async () => {
-      sandbox.stub(globalVariables, "core").value(new MockCore());
-      sandbox.stub(vsc_ui, "VS_CODE_UI").value({
-        selectOption: () =>
-          Promise.resolve(err(new UserError("source", "errorName", "errorMessage"))),
-      });
-
-      const result = await handlers.manageCollaboratorHandler();
-      chai.expect(result.isErr()).equals(true);
     });
   });
 
@@ -1753,307 +1517,6 @@ describe("autoOpenProjectHandler", () => {
   afterEach(() => {
     sandbox.restore();
   });
-  it("opens walk through", async () => {
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openWalkThrough") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    const executeCommandFunc = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendTelemetryStub.calledOnce);
-    chai.assert.isTrue(executeCommandFunc.calledOnce);
-  });
-
-  it("opens walk through if workspace Uri exists", async () => {
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openWalkThrough") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    const globalStateUpdateStub = sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.parse("test"));
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    const executeCommandFunc = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendTelemetryStub.calledOnce);
-    chai.assert.isTrue(executeCommandFunc.calledOnce);
-    chai.assert.isTrue(globalStateUpdateStub.calledTwice);
-  });
-
-  it("opens README", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(manifestUtils, "_readAppManifest").resolves(ok({} as any));
-    sandbox.stub(ManifestUtil, "parseCommonProperties").resolves({ isCopilotPlugin: false });
-    sandbox.stub(globalState, "globalStateUpdate");
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendTelemetryStub.calledOnce);
-  });
-
-  it("opens sample README", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox.stub(vscode.window, "showInformationMessage");
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openSampleReadMe") {
-        return true;
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(executeCommandStub.calledOnce);
-  });
-
-  it("opens README and show APIE ME warnings successfully", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else if (key === GlobalKey.CreateWarnings) {
-        return JSON.stringify([{ type: "type", content: "content" }]);
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-
-    sandbox.stub(manifestUtils, "_readAppManifest").resolves(
-      ok({
-        name: { short: "short", full: "full" },
-        description: { short: "short", full: "" },
-        composeExtensions: [{ commands: [{ id: "command1" }] }],
-      } as any)
-    );
-    const parseRes = {
-      id: "",
-      version: "",
-      capabilities: [""],
-      manifestVersion: "",
-      isApiME: true,
-      isSPFx: false,
-      isApiMeAAD: false,
-    };
-    const parseManifestStub = sandbox.stub(ManifestUtil, "parseCommonProperties").returns(parseRes);
-    VsCodeLogInstance.outputChannel = {
-      show: () => {},
-      info: () => {},
-    } as unknown as vscode.OutputChannel;
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendTelemetryStub.calledTwice);
-    chai.assert.isTrue(parseManifestStub.called);
-  });
-
-  it("opens README and show copilot plugin warnings successfully", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    sandbox.stub(vscode.window, "showInformationMessage").resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else if (key === GlobalKey.CreateWarnings) {
-        return JSON.stringify([{ type: "type", content: "content" }]);
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(path, "relative").returns("test");
-
-    sandbox.stub(manifestUtils, "_readAppManifest").resolves(
-      ok({
-        name: { short: "short", full: "full" },
-        description: { short: "short", full: "" },
-        copilotExtensions: { plugins: [{ file: "ai-plugin.json", id: "plugin1" }] },
-      } as any)
-    );
-    const parseRes = {
-      id: "",
-      version: "",
-      capabilities: ["plugin"],
-      manifestVersion: "",
-      isApiME: false,
-      isSPFx: false,
-      isApiMeAAD: false,
-    };
-    const parseManifestStub = sandbox.stub(ManifestUtil, "parseCommonProperties").returns(parseRes);
-    const getApiSpecStub = sandbox
-      .stub(pluginManifestUtils, "getApiSpecFilePathFromTeamsManifest")
-      .resolves(ok(["test"]));
-    VsCodeLogInstance.outputChannel = {
-      show: () => {},
-      info: () => {},
-    } as unknown as vscode.OutputChannel;
-    const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendTelemetryStub.calledTwice);
-    chai.assert.isTrue(parseManifestStub.called);
-    chai.assert.isTrue(getApiSpecStub.called);
-  });
-  it("skip show warnings if parsing error", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else if (key === GlobalKey.CreateWarnings) {
-        return "string";
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    const sendErrorTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryErrorEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendErrorTelemetryStub.called);
-  });
-
-  it("skip show warnings if cannot get manifest", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else if (key === GlobalKey.CreateWarnings) {
-        return "string";
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox
-      .stub(manifestUtils, "_readAppManifest")
-      .resolves(err(new UserError("source", "name", "", "")));
-
-    const sendErrorTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryErrorEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendErrorTelemetryStub.called);
-  });
-
-  it("skip show warnings if get plugin api spec error", async () => {
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    sandbox.stub(globalVariables, "isTeamsFxProject").resolves(false);
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .resolves(undefined);
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "fx-extension.openReadMe") {
-        return vscode.Uri.file("test").fsPath;
-      } else if (key === GlobalKey.CreateWarnings) {
-        return JSON.stringify([{ type: "type", content: "content" }]);
-      } else {
-        return "";
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-
-    sandbox.stub(manifestUtils, "_readAppManifest").resolves(
-      ok({
-        name: { short: "short", full: "full" },
-        description: { short: "short", full: "" },
-        copilotExtensions: { plugins: [{ file: "ai-plugin.json", id: "plugin1" }] },
-      } as any)
-    );
-    const parseRes = {
-      id: "",
-      version: "",
-      capabilities: ["plugin"],
-      manifestVersion: "",
-      isApiME: false,
-      isSPFx: false,
-      isApiBasedMe: true,
-      isApiMeAAD: false,
-    };
-    sandbox.stub(ManifestUtil, "parseCommonProperties").returns(parseRes);
-    const getApiSpecStub = sandbox
-      .stub(pluginManifestUtils, "getApiSpecFilePathFromTeamsManifest")
-      .resolves(err(new SystemError("test", "test", "", "")));
-    VsCodeLogInstance.outputChannel = {
-      show: () => {},
-      info: () => {},
-    } as unknown as vscode.OutputChannel;
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    const sendErrorTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryErrorEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(sendErrorTelemetryStub.called);
-    chai.assert.equal(
-      sendErrorTelemetryStub.args[0][0],
-      TelemetryEvent.ShowScaffoldingWarningSummaryError
-    );
-    chai.assert.isTrue(getApiSpecStub.called);
-  });
-
-  it("auto install dependency", async () => {
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "teamsToolkit:autoInstallDependency") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    const globalStateStub = sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(vsc_ui, "VS_CODE_UI").value(new VsCodeUI(<vscode.ExtensionContext>{}));
-    const runCommandStub = sandbox.stub(vsc_ui.VS_CODE_UI, "runCommand");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-
-    await handlers.autoOpenProjectHandler();
-
-    chai.assert.isTrue(globalStateStub.calledWith("teamsToolkit:autoInstallDependency", false));
-    chai.assert.isTrue(runCommandStub.calledOnce);
-  });
 
   it("openFolderHandler()", async () => {
     const sendTelemetryStub = sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
@@ -2074,215 +1537,6 @@ describe("autoOpenProjectHandler", () => {
 
     chai.assert.isTrue(sendTelemetryStub.called);
     chai.assert.isTrue(result.isErr());
-  });
-
-  it("openSampleReadmeHandler() - trigger from walkthrough", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.openSampleReadmeHandler(["WalkThrough"]);
-
-    chai.assert.isTrue(executeCommandStub.calledOnce);
-  });
-
-  it("showLocalDebugMessage() - has local env", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("win32");
-    sandbox.stub(fs, "pathExists").resolves(true);
-    const runLocalDebug = sandbox.stub(runIconHandlers, "selectAndDebug").resolves(ok(null));
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Debug",
-            run: (options as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.calledOnce);
-    chai.assert.isTrue(runLocalDebug.called);
-  });
-
-  it("showLocalDebugMessage() - local env and non windows", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("linux");
-    sandbox.stub(fs, "pathExists").resolves(true);
-    const runLocalDebug = sandbox.stub(runIconHandlers, "selectAndDebug").resolves(ok(null));
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Not Debug",
-            run: (options as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.calledOnce);
-    chai.assert.isFalse(runLocalDebug.called);
-  });
-
-  it("showLocalDebugMessage() - has local env and not click debug", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("win32");
-    sandbox.stub(fs, "pathExists").resolves(true);
-    const runLocalDebug = sandbox.stub(runIconHandlers, "selectAndDebug").resolves(ok(null));
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve(undefined);
-        }
-      );
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.calledOnce);
-    chai.assert.isFalse(runLocalDebug.called);
-  });
-
-  it("showLocalDebugMessage() - no local env", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("win32");
-    sandbox.stub(fs, "pathExists").resolves(false);
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Provision",
-            run: (options as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.called);
-    chai.assert.isTrue(executeCommandStub.called);
-  });
-
-  it("showLocalDebugMessage() - no local env and non windows", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(appDefinitionUtils, "getAppName").resolves("");
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("linux");
-    sandbox.stub(fs, "pathExists").resolves(false);
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve({
-            title: "Not provision",
-            run: (options as any).run,
-          } as vscode.MessageItem);
-        }
-      );
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.called);
-    chai.assert.isTrue(executeCommandStub.notCalled);
-  });
-
-  it("showLocalDebugMessage() - no local env and not click provision", async () => {
-    sandbox.stub(vscode.workspace, "workspaceFolders").value([{ uri: vscode.Uri.file("test") }]);
-    sandbox.stub(vscode.workspace, "openTextDocument");
-    sandbox.stub(process, "platform").value("win32");
-    sandbox.stub(fs, "pathExists").resolves(false);
-
-    sandbox.stub(globalState, "globalStateGet").callsFake(async (key: string) => {
-      if (key === "ShowLocalDebugMessage") {
-        return true;
-      } else {
-        return false;
-      }
-    });
-    sandbox.stub(globalState, "globalStateUpdate");
-    sandbox.stub(ExtTelemetry, "sendTelemetryEvent");
-    sandbox.stub(globalVariables, "workspaceUri").value(vscode.Uri.file("test"));
-    const showMessageStub = sandbox
-      .stub(vscode.window, "showInformationMessage")
-      .callsFake(
-        (title: string, options: vscode.MessageOptions, ...items: vscode.MessageItem[]) => {
-          return Promise.resolve(undefined);
-        }
-      );
-    const executeCommandStub = sandbox.stub(vscode.commands, "executeCommand");
-
-    await handlers.showLocalDebugMessage();
-
-    chai.assert.isTrue(showMessageStub.called);
-    chai.assert.isFalse(executeCommandStub.called);
   });
 
   it("installAdaptiveCardExt()", async () => {
